Reuse MenuList and ClearIcon components
