+++ webapp/src/components/backstage/automation/assign_owner_selector.tsx
@@ -1,9 +1,7 @@
 import React, {useState, useEffect} from 'react';
 import {useSelector} from 'react-redux';
 
-import ReactSelect, {ControlProps, MenuListComponentProps} from 'react-select';
-
-import {Scrollbars} from 'react-custom-scrollbars';
+import ReactSelect, {ControlProps} from 'react-select';
 
 import styled from 'styled-components';
 import {ActionFunc} from 'mattermost-redux/types/actions';
@@ -12,7 +10,8 @@ import {UserProfile} from 'mattermost-redux/types/users';
 import {getUser} from 'mattermost-redux/selectors/entities/users';
 
 import Profile from 'src/components/profile/profile';
-import ClearIcon from 'src/components/assets/icons/clear_icon';
+import ClearIndicator from 'src/components/backstage/automation/clear_indicator';
+import MenuList from 'src/components/backstage/automation/menu_list';
 
 interface Props {
     ownerID: string;
@@ -181,57 +180,3 @@ const StyledReactSelect = styled(ReactSelect)`
         color: rgba(var(--center-channel-color-rgb), 0.56);
     }
 `;
-
-const MenuListWrapper = styled.div`
-    background-color: var(--center-channel-bg);
-    border: 1px solid rgba(var(--center-channel-color-rgb), 0.16);
-    border-radius: 4px;
-
-    max-height: 280px;
-`;
-
-const MenuHeaderHeight = 44;
-
-const MenuHeader = styled.div`
-    height: ${MenuHeaderHeight}px;
-    padding: 16px 0 12px 14px;
-    font-size: 14px;
-    font-weight: 600;
-    border-bottom: 1px solid rgba(var(--center-channel-color-rgb), 0.08);
-    line-height: 16px;
-`;
-
-const StyledScrollbars = styled(Scrollbars)`
-    height: ${300 - MenuHeaderHeight}px;
-`;
-
-const ThumbVertical = styled.div`
-    background-color: rgba(var(--center-channel-color-rgb), 0.24);
-    border-radius: 2px;
-    width: 4px;
-    min-height: 45px;
-    margin-left: -2px;
-    margin-top: 6px;
-`;
-
-const MenuList = (props: MenuListComponentProps<UserProfile>) => {
-    return (
-        <MenuListWrapper>
-            <MenuHeader>{'Assign Owner'}</MenuHeader>
-            <StyledScrollbars
-                autoHeight={true}
-                renderThumbVertical={({style, ...thumbProps}) => <ThumbVertical {...thumbProps}/>}
-            >
-                {props.children}
-            </StyledScrollbars>
-        </MenuListWrapper>
-    );
-};
-
-const ClearIndicator = ({clearValue}: {clearValue: () => void}) => {
-    return (
-        <div onClick={clearValue}>
-            <ClearIcon/>
-        </div>
-    );
-};
