+++ webapp/src/components/backstage/automation/broadcast.tsx
@@ -3,16 +3,14 @@
 
 import React from 'react';
 
-import {MenuListComponentProps} from 'react-select';
-import {Scrollbars} from 'react-custom-scrollbars';
 import styled from 'styled-components';
-
-import {Channel} from 'mattermost-redux/types/channels';
 
 import {AutomationHeader, AutomationTitle, SelectorWrapper} from 'src/components/backstage/automation/styles';
 import {Toggle} from 'src/components/backstage/automation/toggle';
 import ChannelSelector from 'src/components/backstage/channel_selector';
 import ClearIcon from 'src/components/assets/icons/clear_icon';
+import ClearIndicator from 'src/components/backstage/automation/clear_indicator';
+import MenuList from 'src/components/backstage/automation/menu_list';
 
 interface Props {
     enabled: boolean;
@@ -67,58 +65,6 @@ const StyledChannelSelector = styled(ChannelSelector)`
         }
     }
 `;
-
-const MenuListWrapper = styled.div`
-    background-color: var(--center-channel-bg);
-    border: 1px solid rgba(var(--center-channel-color-rgb), 0.16);
-    border-radius: 4px;
-
-    max-height: 280px;
-`;
-
-const MenuHeaderHeight = 44;
-
-const MenuHeader = styled.div`
-    height: ${MenuHeaderHeight}px;
-    padding: 16px 0 12px 14px;
-    font-size: 14px;
-    font-weight: 600;
-    border-bottom: 1px solid rgba(var(--center-channel-color-rgb), 0.08);
-    line-height: 16px;
-`;
-
-const StyledScrollbars = styled(Scrollbars)`
-    height: ${300 - MenuHeaderHeight}px;
-`;
-
-const ThumbVertical = styled.div`
-    background-color: rgba(var(--center-channel-color-rgb), 0.24);
-    border-radius: 2px;
-    width: 4px;
-    min-height: 45px;
-    margin-left: -2px;
-    margin-top: 6px;
-`;
-
-const MenuList = (props: MenuListComponentProps<Channel>) => {
-    return (
-        <MenuListWrapper>
-            <MenuHeader>{'Announcement channel'}</MenuHeader>
-            <StyledScrollbars
-                autoHeight={true}
-                renderThumbVertical={({style, ...thumbProps}) => <ThumbVertical {...thumbProps}/>}
-            >
-                {props.children}
-            </StyledScrollbars>
-        </MenuListWrapper>
-    );
-};
-
-const ClearIndicator = ({clearValue}: {clearValue: () => void}) => (
-    <div onClick={clearValue}>
-        <ClearIcon/>
-    </div>
-);
 
 interface MultiValueRemoveProps {
     innerProps: {
