+++ webapp/src/components/backstage/automation/categorize_playbook_run.tsx
@@ -3,17 +3,14 @@
 
 import React from 'react';
 
-import {MenuListComponentProps} from 'react-select';
-import {Scrollbars} from 'react-custom-scrollbars';
 import styled from 'styled-components';
-
-import {ChannelCategory} from 'mattermost-redux/types/channel_categories';
 
 import {AutomationHeader, AutomationTitle, SelectorWrapper} from 'src/components/backstage/automation/styles';
 import {Toggle} from 'src/components/backstage/automation/toggle';
+import ClearIndicator from 'src/components/backstage/automation/clear_indicator';
+import MenuList from 'src/components/backstage/automation/menu_list';
 
 import CategorySelector from 'src/components/backstage/category_selector';
-import ClearIcon from 'src/components/assets/icons/clear_icon';
 
 interface Props {
     enabled: boolean;
@@ -66,55 +63,3 @@ const StyledCategorySelector = styled(CategorySelector)`
         }
     }
 `;
-
-const MenuListWrapper = styled.div`
-    background-color: var(--center-channel-bg);
-    border: 1px solid rgba(var(--center-channel-color-rgb), 0.16);
-    border-radius: 4px;
-
-    max-height: 280px;
-`;
-
-const MenuHeaderHeight = 44;
-
-const MenuHeader = styled.div`
-    height: ${MenuHeaderHeight}px;
-    padding: 16px 0 12px 14px;
-    font-size: 14px;
-    font-weight: 600;
-    border-bottom: 1px solid rgba(var(--center-channel-color-rgb), 0.08);
-    line-height: 16px;
-`;
-
-const StyledScrollbars = styled(Scrollbars)`
-    height: ${300 - MenuHeaderHeight}px;
-`;
-
-const ThumbVertical = styled.div`
-    background-color: rgba(var(--center-channel-color-rgb), 0.24);
-    border-radius: 2px;
-    width: 4px;
-    min-height: 45px;
-    margin-left: -2px;
-    margin-top: 6px;
-`;
-
-const MenuList = (props: MenuListComponentProps<ChannelCategory>) => {
-    return (
-        <MenuListWrapper>
-            <MenuHeader>{'Select a category'}</MenuHeader>
-            <StyledScrollbars
-                autoHeight={true}
-                renderThumbVertical={({style, ...thumbProps}) => <ThumbVertical {...thumbProps}/>}
-            >
-                {props.children}
-            </StyledScrollbars>
-        </MenuListWrapper>
-    );
-};
-
-const ClearIndicator = ({clearValue}: {clearValue: () => void}) => (
-    <div onClick={clearValue}>
-        <ClearIcon/>
-    </div>
-);
