+++ webapp/src/components/backstage/automation/clear_indicator.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import ClearIcon from 'src/components/assets/icons/clear_icon';
+
+const ClearIndicator = ({clearValue}: {clearValue: () => void}) => (
+    <div onClick={clearValue}>
+        <ClearIcon/>
+    </div>
+);
+
+export default ClearIndicator;
