+++ webapp/src/components/backstage/automation/invite_users_selector.tsx
@@ -4,9 +4,7 @@ import ReactSelect, {
     GroupType,
     ControlProps,
     OptionsType,
-    MenuListComponentProps,
 } from 'react-select';
-import {Scrollbars} from 'react-custom-scrollbars';
 
 import styled from 'styled-components';
 import {ActionFunc} from 'mattermost-redux/types/actions';
@@ -16,6 +14,8 @@ import {getUser} from 'mattermost-redux/selectors/entities/users';
 
 import Profile from 'src/components/profile/profile';
 import {useEnsureProfiles} from 'src/hooks';
+
+import MenuList from 'src/components/backstage/automation/menu_list';
 
 interface Props {
     userIds: string[];
@@ -282,49 +282,3 @@ const StyledReactSelect = styled(ReactSelect)`
         color: rgba(var(--center-channel-color-rgb), 0.56);
     }
 `;
-
-const MenuListWrapper = styled.div`
-    background-color: var(--center-channel-bg);
-    border: 1px solid rgba(var(--center-channel-color-rgb), 0.16);
-    border-radius: 4px;
-
-    max-height: 280px;
-`;
-
-const MenuHeaderHeight = 44;
-
-const MenuHeader = styled.div`
-    height: ${MenuHeaderHeight}px;
-    padding: 16px 0 12px 14px;
-    font-size: 14px;
-    font-weight: 600;
-    border-bottom: 1px solid rgba(var(--center-channel-color-rgb), 0.08);
-    line-height: 16px;
-`;
-
-const StyledScrollbars = styled(Scrollbars)`
-    height: ${300 - MenuHeaderHeight}px;
-`;
-
-const ThumbVertical = styled.div`
-    background-color: rgba(var(--center-channel-color-rgb), 0.24);
-    border-radius: 2px;
-    width: 4px;
-    min-height: 45px;
-    margin-left: -2px;
-    margin-top: 6px;
-`;
-
-const MenuList = (props: MenuListComponentProps<UserProfile>) => {
-    return (
-        <MenuListWrapper>
-            <MenuHeader>{'Invite Members'}</MenuHeader>
-            <StyledScrollbars
-                autoHeight={true}
-                renderThumbVertical={({style, ...thumbProps}) => <ThumbVertical {...thumbProps}/>}
-            >
-                {props.children}
-            </StyledScrollbars>
-        </MenuListWrapper>
-    );
-};
