+++ webapp/src/components/backstage/automation/menu_list.tsx
@@ -1,0 +1,53 @@
+import React from 'react';
+
+import styled from 'styled-components';
+import {Scrollbars} from 'react-custom-scrollbars';
+import {MenuListComponentProps, OptionTypeBase} from 'react-select';
+
+const MenuListWrapper = styled.div`
+    background-color: var(--center-channel-bg);
+    border: 1px solid rgba(var(--center-channel-color-rgb), 0.16);
+    border-radius: 4px;
+
+    max-height: 280px;
+`;
+
+const MenuHeaderHeight = 44;
+
+const MenuHeader = styled.div`
+    height: ${MenuHeaderHeight}px;
+    padding: 16px 0 12px 14px;
+    font-size: 14px;
+    font-weight: 600;
+    border-bottom: 1px solid rgba(var(--center-channel-color-rgb), 0.08);
+    line-height: 16px;
+`;
+
+const StyledScrollbars = styled(Scrollbars)`
+    height: ${300 - MenuHeaderHeight}px;
+`;
+
+const ThumbVertical = styled.div`
+    background-color: rgba(var(--center-channel-color-rgb), 0.24);
+    border-radius: 2px;
+    width: 4px;
+    min-height: 45px;
+    margin-left: -2px;
+    margin-top: 6px;
+`;
+
+const MenuList = <T extends OptionTypeBase>(props: MenuListComponentProps<T>) => {
+    return (
+        <MenuListWrapper>
+            {props.selectProps.placeholder && <MenuHeader>{props.selectProps.placeholder}</MenuHeader>}
+            <StyledScrollbars
+                autoHeight={true}
+                renderThumbVertical={({style, ...thumbProps}) => <ThumbVertical {...thumbProps}/>}
+            >
+                {props.children}
+            </StyledScrollbars>
+        </MenuListWrapper>
+    );
+};
+
+export default MenuList;
